main.py

import logging
import requests
from bs4 import BeautifulSoup
from telegram import Update
from telegram.ext import Updater, CommandHandler, CallbackContext

# Вставьте сюда ваш URL для проверки
URL = "https://example.com"

def start(update: Update, context: CallbackContext):
    update.message.reply_text("Привет! Я бот для отслеживания терминов.")

def check_appointments(update: Update, context: CallbackContext):
    try:
        response = requests.get(URL)
        soup = BeautifulSoup(response.text, 'html.parser')
        # Здесь добавьте ваш парсинг
        update.message.reply_text("Проверка завершена. Терминов нет.")
    except Exception as e:
        update.message.reply_text(f"Ошибка при проверке: {e}")

def main():
    updater = Updater("ВАШ_ТОКЕН", use_context=True)
    dp = updater.dispatcher
    dp.add_handler(CommandHandler("start", start))
    dp.add_handler(CommandHandler("check", check_appointments))
    updater.start_polling()
    updater.idle()

if __name__ == '__main__':
    main()
